Map Offline => Off
When the Litter Robot is off, sometimes the API reports "OFFLINE" as the
status. It's unclear to me what the pattern really is or if there is
anything meaningful to be gleaned from the Offline status.

For the moment, I'm going to treat it as Off -- since that's the only
time it seems to happen.